// src/components/JobCard/JobCard.types.ts
export type CardDataType = {
    jdUid: string;
    jdLink: string;
    jobDetailsFromCompany: string;
    maxJdSalary: number;
    minJdSalary: number;
    salaryCurrencyCode: string;
    location: string;
    minExp?: number | null;
    maxExp?: number | null;
    jobRole: string;
    companyName: string;
    logoUrl: string;
  };